Add: asm> New parameter type "Trigger"
"Trigger" is a table that contains two fields: type, state

To be honest, asm will recognize table that have `type` field with
value "trigger" as trigger parameter. It is based on unity's trigger
parameter, which will be "true" only for a little time.

When you `Update` that parameter, it will be activated only for that
Update.
To reference trigger parameters from transition triggers, check
`<parameter>.<your_trigger_name>.state`. It'll be true when trigger is
activated.